Move some functions to the helpers module
