+++ src/tespy/tools/document_models.py
@@ -10,8 +10,6 @@ SPDX-License-Identifier: MIT
 """
 import os
 import sys
-from collections.abc import Mapping
-from copy import deepcopy
 from datetime import date
 
 import CoolProp as CP
@@ -77,20 +75,6 @@ def document_model(nw, path='report', filename='report.tex', fmt={}):
     with open(path + filename, 'w') as f:
         f.write(latex)
         f.close()
-
-
-def merge_dicts(dict1, dict2):
-    """Return a new dictionary by merging two dictionaries recursively."""
-
-    result = deepcopy(dict1)
-
-    for key, value in dict2.items():
-        if isinstance(value, Mapping):
-            result[key] = merge_dicts(result.get(key, {}), value)
-        else:
-            result[key] = deepcopy(dict2[key])
-
-    return result
 
 
 def set_defaults(nw):
