+++ src/tespy/tools/helpers.py
@@ -14,11 +14,48 @@ import os
 import CoolProp as CP
 import numpy as np
 
+from collections import OrderedDict
+from collections.abc import Mapping
+from copy import deepcopy
+
 from tespy.tools.global_vars import err
 from tespy.tools.global_vars import fluid_property_data
 from tespy.tools.global_vars import molar_masses
 
-# %%
+
+def merge_dicts(dict1, dict2):
+    """Return a new dictionary by merging two dictionaries recursively."""
+
+    result = deepcopy(dict1)
+
+    for key, value in dict2.items():
+        if isinstance(value, Mapping):
+            result[key] = merge_dicts(result.get(key, {}), value)
+        else:
+            result[key] = deepcopy(dict2[key])
+
+    return result
+
+
+def nested_OrderedDict(dictionary):
+    """Create a nested OrderedDict from a nested dict.
+
+    Parameters
+    ----------
+    dictionary : dict
+        Nested dict.
+
+    Returns
+    -------
+    dictionary : collections.OrderedDict
+        Nested OrderedDict.
+    """
+    dictionary = OrderedDict(dictionary)
+    for key, value in dictionary.items():
+        if isinstance(value, dict):
+            dictionary[key] = nested_OrderedDict(value)
+
+    return dictionary
 
 
 class TESPyNetworkError(Exception):
